+++ src/components/Navbar/constants.tsx
@@ -1,10 +1,61 @@
-import { Trans } from '@lingui/macro'
+import { DownOutlined, UpOutlined } from '@ant-design/icons'
+import { t, Trans } from '@lingui/macro'
+import { Dropdown } from 'antd'
+import Link from 'next/link'
+import { CSSProperties } from 'react'
+import Logo from './Logo'
+import { navMenuItemStyles } from './navStyles'
 
 type ResourceItem = {
   label: JSX.Element
   key: string
 }
+const menuItemProps = {
+  style: navMenuItemStyles,
+  className: 'nav-menu-item hover-opacity',
+}
 
+const externalMenuLinkProps = {
+  ...menuItemProps,
+  target: '_blank',
+  rel: 'noopener noreferrer',
+}
+
+export const DesktopDropDown = ({
+  resourcesMenu,
+  resourcesOpen,
+  setResourcesOpen,
+  dropdownIconStyle,
+}: {
+  resourcesMenu: JSX.Element
+  resourcesOpen: boolean
+  setResourcesOpen: (resource: boolean) => void
+  dropdownIconStyle: CSSProperties
+}): JSX.Element => {
+  return (
+    <Dropdown
+      overlay={resourcesMenu}
+      overlayStyle={{ padding: 0 }}
+      visible={resourcesOpen}
+    >
+      <div
+        className="nav-menu-item hover-opacity"
+        onClick={e => {
+          setResourcesOpen(!resourcesOpen)
+          e.stopPropagation()
+        }}
+        style={{ ...navMenuItemStyles }}
+      >
+        <Trans>Resources</Trans>
+        {resourcesOpen ? (
+          <UpOutlined style={dropdownIconStyle} />
+        ) : (
+          <DownOutlined style={dropdownIconStyle} />
+        )}
+      </div>
+    </Dropdown>
+  )
+}
 export const resourcesMenuItems = (mobile?: boolean): ResourceItem[] => {
   const linkStyle = {
     className: 'nav-dropdown-item',
@@ -50,3 +101,61 @@ export const resourcesMenuItems = (mobile?: boolean): ResourceItem[] => {
     },
   ]
 }
+
+export const desktopMenuItems = ({
+  resourcesMenu,
+  resourcesOpen,
+  setResourcesOpen,
+  dropdownIconStyle,
+}: {
+  resourcesMenu: JSX.Element
+  resourcesOpen: boolean
+  setResourcesOpen: (resource: boolean) => void
+  dropdownIconStyle: CSSProperties
+}) => [
+  {
+    key: 'index',
+    label: (
+      <Link href="/">
+        <a style={{ display: 'inline-block' }}>{<Logo />}</a>
+      </Link>
+    ),
+  },
+  {
+    key: 'projects',
+    label: (
+      <Link href="/projects">
+        <a {...menuItemProps}>{t`Projects`}</a>
+      </Link>
+    ),
+  },
+  {
+    key: 'docs',
+    label: (
+      <Link href="https://info.juicebox.money/">
+        <a {...externalMenuLinkProps}>{t`Docs`}</a>
+      </Link>
+    ),
+  },
+  {
+    key: 'blog',
+    label: (
+      <Link href="https://info.juicebox.money/blog">
+        <a {...externalMenuLinkProps}>{t`Blog`}</a>
+      </Link>
+    ),
+  },
+  {
+    key: 'resources',
+    label: (
+      <DesktopDropDown
+        {...{
+          resourcesMenu,
+          setResourcesOpen,
+          resourcesOpen,
+          dropdownIconStyle,
+        }}
+      />
+    ),
+  },
+]
